Move seven classes from study-api to specimen-main
