+++ specimen/src/org/labkey/specimen/SpecimenManager.java
@@ -26,8 +26,8 @@ import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
-import org.labkey.api.specimen.model.AdditiveType;
-import org.labkey.api.specimen.model.DerivativeType;
+import org.labkey.specimen.model.AdditiveType;
+import org.labkey.specimen.model.DerivativeType;
 import org.labkey.api.specimen.model.ExtendedSpecimenRequestView;
 import org.labkey.api.specimen.model.SpecimenComment;
 import org.labkey.api.study.Cohort;
