+++ specimen/src/org/labkey/specimen/actions/SpecimenApiController.java
@@ -39,8 +39,8 @@ import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.importer.RequestabilityManager;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
-import org.labkey.api.specimen.model.AdditiveType;
-import org.labkey.api.specimen.model.DerivativeType;
+import org.labkey.specimen.model.AdditiveType;
+import org.labkey.specimen.model.DerivativeType;
 import org.labkey.api.specimen.model.PrimaryType;
 import org.labkey.api.specimen.model.SpecimenTypeSummary;
 import org.labkey.api.specimen.requirements.SpecimenRequest;
