+++ specimen/src/org/labkey/specimen/actions/SpecimenController.java
@@ -71,8 +71,8 @@ import org.labkey.api.specimen.model.ExtendedSpecimenRequestView;
 import org.labkey.api.specimen.model.SpecimenComment;
 import org.labkey.api.specimen.model.SpecimenRequestActor;
 import org.labkey.api.specimen.model.SpecimenRequestEvent;
-import org.labkey.api.specimen.notifications.DefaultRequestNotification;
-import org.labkey.api.specimen.notifications.NotificationRecipientSet;
+import org.labkey.specimen.notifications.DefaultRequestNotification;
+import org.labkey.specimen.notifications.NotificationRecipientSet;
 import org.labkey.api.specimen.query.SpecimenQueryView;
 import org.labkey.api.specimen.requirements.SpecimenRequest;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirement;
@@ -87,8 +87,8 @@ import org.labkey.api.specimen.settings.RepositorySettings;
 import org.labkey.api.specimen.settings.RequestNotificationSettings;
 import org.labkey.api.specimen.settings.SettingsManager;
 import org.labkey.api.specimen.settings.StatusSettings;
-import org.labkey.api.specimen.view.NotificationBean;
-import org.labkey.api.specimen.view.SpecimenRequestNotificationEmailTemplate;
+import org.labkey.specimen.view.NotificationBean;
+import org.labkey.specimen.view.SpecimenRequestNotificationEmailTemplate;
 import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.Dataset;
 import org.labkey.api.study.MapArrayExcelWriter;
