+++ specimen/src/org/labkey/specimen/model/DerivativeType.java
@@ -13,31 +13,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.labkey.api.specimen.model;
+package org.labkey.specimen.model;
 
 import org.labkey.api.data.Container;
+import org.labkey.api.specimen.model.DerivativeTypeDomainKind;
 import org.labkey.api.study.AbstractStudyCachable;
 
 import java.util.Map;
 
-public class AdditiveType extends AbstractStudyCachable<AdditiveType>
+/*
+ * User: brittp
+ * Date: Dec 18, 2008
+ * Time: 12:48:03 PM
+ */
+
+public class DerivativeType extends AbstractStudyCachable<DerivativeType>
 {
     private long _rowId; // serial NOT NULL,
     private Container _container; // entityid NOT NULL,
-    private String _ldmsAdditiveCode; // character varying(5),
-    private String _labwareAdditiveCode; // character varying(5),
-    private String _additive; // character varying(100),
+    private String _ldmsDerivativeCode; // character varying(5),
+    private String _labwareDerivativeCode; // character varying(5),
+    private String _derivative; // character varying(100),
     private int _externalId; // integer NOT NULL default 0
 
-    public AdditiveType(Container container, Map<String, Object> map)
+    public DerivativeType(Container container, Map<String, Object> map)
     {
         super();
-        _rowId = (int) map.get(AdditiveTypeDomainKind.ROWID);
+        _rowId = (int) map.get(DerivativeTypeDomainKind.ROWID);
         _container = container;
-        _ldmsAdditiveCode = (String) map.get(AdditiveTypeDomainKind.LDMSADDITIVECODE);
-        _labwareAdditiveCode = (String) map.get(AdditiveTypeDomainKind.LABWAREADDITIVECODE);
-        _additive = (String) map.get(AdditiveTypeDomainKind.ADDITIVE);
-        _externalId = (int) map.get(AdditiveTypeDomainKind.EXTERNALID);
+        _ldmsDerivativeCode = (String) map.get(DerivativeTypeDomainKind.LDMSDERIVATIVECODE);
+        _labwareDerivativeCode = (String) map.get(DerivativeTypeDomainKind.LABWAREDERIVATIVECODE);
+        _derivative = (String) map.get(DerivativeTypeDomainKind.DERIVATIVE);
+        _externalId = (int) map.get(DerivativeTypeDomainKind.EXTERNALID);
     }
 
     @Override
@@ -78,33 +85,33 @@ public class AdditiveType extends AbstractStudyCachable<AdditiveType>
         _container = container;
     }
 
-    public String getLdmsAdditiveCode()
+    public String getLdmsDerivativeCode()
     {
-        return _ldmsAdditiveCode;
+        return _ldmsDerivativeCode;
     }
 
-    public void setLdmsAdditiveCode(String ldmsAdditiveCode)
+    public void setLdmsDerivativeCode(String ldmsDerivativeCode)
     {
-        _ldmsAdditiveCode = ldmsAdditiveCode;
+        _ldmsDerivativeCode = ldmsDerivativeCode;
     }
 
-    public String getLabwareAdditiveCode()
+    public String getLabwareDerivativeCode()
     {
-        return _labwareAdditiveCode;
+        return _labwareDerivativeCode;
     }
 
-    public void setLabwareAdditiveCode(String labwareAdditiveCode)
+    public void setLabwareDerivativeCode(String labwareDerivativeCode)
     {
-        _labwareAdditiveCode = labwareAdditiveCode;
+        _labwareDerivativeCode = labwareDerivativeCode;
     }
 
-    public String getAdditive()
+    public String getDerivative()
     {
-        return _additive;
+        return _derivative;
     }
 
-    public void setAdditive(String additive)
+    public void setDerivative(String derivative)
     {
-        _additive = additive;
+        _derivative = derivative;
     }
 }
