+++ specimen/src/org/labkey/specimen/notifications/ActorNotificationRecipientSet.java
@@ -21,7 +21,6 @@ import org.labkey.api.security.User;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.SpecimenRequestActor;
-import org.labkey.api.specimen.notifications.NotificationRecipientSet;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.HtmlStringBuilder;
