+++ study/api-src/org/labkey/api/specimen/SpecimenMigrationService.java
@@ -6,7 +6,6 @@ import org.labkey.api.data.TableInfo;
 import org.labkey.api.pipeline.PipelineJob;
 import org.labkey.api.pipeline.PipelineJobException;
 import org.labkey.api.services.ServiceRegistry;
-import org.labkey.api.specimen.model.SpecimenRequestEvent;
 import org.labkey.api.study.importer.SimpleStudyImportContext;
 import org.labkey.api.view.ActionURL;
 
@@ -26,11 +25,12 @@ public interface SpecimenMigrationService
         ServiceRegistry.get().registerService(SpecimenMigrationService.class, impl);
     }
 
+    String SPECIMENS_ARCHIVE_TYPE = "Specimens";
+
     ActionURL getBeginURL(Container c);
     ActionURL getInsertSpecimenQueryRowURL(Container c, String schemaName, TableInfo table);
     ActionURL getSelectedSpecimensURL(Container c);
     ActionURL getSpecimenEventsURL(Container c, ActionURL returnUrl);
-    ActionURL getSpecimenRequestEventDownloadURL(SpecimenRequestEvent event, String name);
     ActionURL getSpecimensURL(Container c);
     ActionURL getUpdateSpecimenQueryRowURL(Container c, String schemaName, TableInfo table);
