--- study/api-src/org/labkey/api/specimen/writer/SpecimenArchiveDataTypes.java
@@ -1,13 +1,0 @@
-package org.labkey.api.specimen.writer;
-
-public class SpecimenArchiveDataTypes
-{
-    public static final String SCHEMA_FILENAME = "specimens_metadata.xml";
-
-    private SpecimenArchiveDataTypes()
-    {
-    }
-
-    public static final String SPECIMEN_SETTINGS =  "Specimen Settings";
-    public static final String SPECIMENS = "Specimens";
-}
