+++ study/resources/schemas/dbscripts/postgresql/study-22.000-22.001.sql
@@ -1,0 +1,1 @@
+SELECT core.executeJavaUpgradeCode('moveSpecimenTemplatePropertiesAgain');
