+++ study/resources/schemas/dbscripts/sqlserver/study-22.000-22.001.sql
@@ -1,0 +1,1 @@
+EXEC core.executeJavaUpgradeCode 'moveSpecimenTemplatePropertiesAgain';
